added option to preserve .o files

+++ testSimplec.py
@@ -12,7 +12,7 @@ test_case_points = 2
 build_points = 0 # points for building. tentative
 
 
-def buildAndTest(submissionpath, sourceTestPath):
+def buildAndTest(submissionpath, sourceTestPath, no_remove):
     
 
     points = 0
@@ -23,10 +23,10 @@ def buildAndTest(submissionpath, sourceTestPath):
 
     testCases = glob.glob(os.path.join(testCasePath, "*.simplec"))
     #print(f"testCases {testCases}")
-
-    for i in glob.glob(os.path.join(submissionpath, "*.o")):
-        if os.path.exists(i):
-            os.remove(i)
+    if not no_remove:
+        for i in glob.glob(os.path.join(submissionpath, "*.o")):
+            if os.path.exists(i):
+                os.remove(i)
     progname = os.path.join(submissionpath, "simplec")
     if os.path.exists(progname):
         os.remove(progname)
@@ -95,6 +95,12 @@ if __name__ == "__main__":
         print("USAGE: path/to/your/repo path/to/the/tests")
         print("example: ./ ../syllabus/projects/tests/proj0/")
         sys.exit()
+    no_remove = False
+    try:
+        temp = sys.argv[3]
+    except:
+        pass
+    if(len(temp) > 0):
+        no_remove = True
 
-
-    buildAndTest(submissionDirectory, sourceTestPath)
+    buildAndTest(submissionDirectory, sourceTestPath, no_remove)
